Split deployment provider state from navigation
